requirements.txt
streamlit
pandas
matplotlib
langchain
langgraph
langchain-openai 
python-dotenv